Added user-controlled payload orientation in editor without limits
Removed AutoPositionVessel as an option to SpawnSpaceManager; is is now an
option of HangarEntrance and HangarStorage and is used solely in editor for
INITIAL optional payload orientation; afterwards the orientation may be changed
by user.

Reimplemented the whole process of storage checks to accommodate new orientation
system.

In-flight the orientation of the docking vessel is preserved to some extent: the
orthogonal spawn rotation is chosen so that it is closest to the original vessel
rotation.

Added HangarMachinery.PayloadFixedInFlight to forbid orientation changes in
flight for singel use hangars and fairings.