ProceduralAdapter: tried to fix the problem with selection of the original part
That is, when a part is added as a symmetry group, the original part cannot be
selected unless it is resized as a part of the group via PAW of another group
member.

In vain, but it stands to reason to first disable the collider, then replace its
mesh and only then enable it again. As well as invalidate all renderer caches of
the part.